Install database in case no cache was found.

+++ tests/profile_setup.sh
@@ -8,8 +8,11 @@ export COLLECTIVEACCESS_HOME="$(dirname $(dirname "$0"))"
 export PATH="$PATH:$COLLECTIVEACCESS_HOME/support/bin"
 
 # Install the testing profile
-"$COLLECTIVEACCESS_HOME"/support/bin/caUtils install --hostname=localhost --setup="tests/setup-tests.php" \
-  --skip-roles --profile-name="$PROFILE" --admin-email=support@collectiveaccess.org
+if test ! -e "$CACHE_DIR/$PROFILE.sql";
+then
+  "$COLLECTIVEACCESS_HOME"/support/bin/caUtils install --hostname=localhost --setup="tests/setup-tests.php" \
+    --skip-roles --profile-name="$PROFILE" --admin-email=support@collectiveaccess.org
+fi
 
 # Export database for later faster import
 sudo mysqldump -uroot $DB_NAME > "$CACHE_DIR/$PROFILE.sql"
